first pass at splitting up API
