+++ Dockerfile-test
@@ -4,5 +4,5 @@ COPY tests/package.json .
 RUN npm install
 COPY tests .
 COPY nodejs-server/helpers/helpers.js tests/helpers.js
-COPY spec.json .
-CMD npm test
+COPY core-spec.json .
+COPY miami-spec.json .
